requirements.txt
mcp>=1.2.0
tweepy>=4.16.0
python-dotenv>=1.2.1
requests>=2.32.5
beautifulsoup4>=4.14.3
